tests: Factor out check_fuse() into a file that can be shared by Flatpak
